+++ src/binance.ts
@@ -23,7 +23,7 @@ declare module 'binance-api-node' {
   export interface SymbolMinNotionalFilter {
     filterType: SymbolFilterType.MIN_NOTIONAL
     minNotional: string
-    notional?: string
+    notional: string
   }
   export interface Symbol<T = OrderType_LT> {
     baseAsset: string